LayoutTests/accessibility/aria-table-content.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
<style type="text/css">
  div.foo table:not(.no-header):before{content:'';}
</style>
<title>Table Anonymous Row</title>
</head>
<body>
<div class="foo" role="grid" id="table">
  <table>
	<thead> <tr> <th>Header</th> </tr> </thead>
	<tbody> <tr> <td> <p>Item 1</p> </td> </tr> </tbody>
  </table>
</div>

<p id="description"></p>
<div id="console"></div>

<script>
description("This tests that in an aria table with CSS that makes a row anonymous, the cells can be accessed.");

if (window.accessibilityController) {
    var table = accessibilityController.accessibleElementById("table");
    var cell00 = table.cellForColumnAndRow(0,0);
    shouldBe("cell00.role", "'AXRole: AXCell'");
    var cell01 = table.cellForColumnAndRow(0,1);
    shouldBe("cell01.role", "'AXRole: AXCell'");
    shouldBeTrue("cell00.isEqual(table.rowAtIndex(0).childAtIndex(0))");
    shouldBeTrue("cell01.isEqual(table.rowAtIndex(1).childAtIndex(0))");
}

</script>

<script src="../resources/js-test-post.js"></script>
</body>
</html>
