LayoutTests/platform/glib/accessibility/aria-table-content-expected.txt

Header
Item 1

This tests that in an aria table with CSS that makes a row anonymous, the cells can be accessed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


cell00.role is AXRole: AXColumnHeader
cell01.role is AXRole: AXCell
PASS cell00.isEqual(table.rowAtIndex(0).childAtIndex(0)) is true
PASS cell01.isEqual(table.rowAtIndex(1).childAtIndex(0)) is true
PASS successfullyParsed is true

TEST COMPLETE
